Serve the input bundle from a url instead of in a script tag.
This prevents things from blowing up if the passed in scripts happen to contain the string "</script>".

This also spits up node integration slightly because it wants a full html page.

While at it I switched to storing the source url in an actual temp file unless needed to be in-directory for node integration.